kubernetes-client-examples/python/requirements.txt
kubernetes~=31.0.0

PyYAML~=6.0.1